Added Homing particles + optimization

+++ ufogame/src/main/java/com/reds2/school/Particle.java
@@ -3,16 +3,16 @@ package com.reds2.school;
 import java.awt.Color;
 
 public class Particle {
-    int stage;
+    double stage;
     Color color;
-    int xV = 0,yV = 0,x,y;
-    Particle(int x, int y,int stage,Color color){
+    double xV = 0,yV = 0,x,y;
+    Particle(double x, double y,double stage,Color color){
         this.stage = stage;
         this.color = color;
         this.x = x;
         this.y = y;
     }
-    Particle(int x, int y,int xV, int yV,int stage,Color color){
+    Particle(double x, double y,double xV, double yV,double stage,Color color){
         this.stage = stage;
         this.color = color;
         this.xV = xV;
